Docker_Project/anonymisation_internal/anonymisation_api/anonymise/requirements.txt
transformers==4.18.0
stanza==1.3.0
torch==1.11.0
protobuf==3.20.1
flask==1.1.4
markupsafe==2.0.1
exrex
estnltk==1.7.0
gunicorn